requirements.txt
openai
faiss-cpu
streamlit
python-dotenv
numpy
requests
beautifulsoup4
tqdm